Title: 최대 2번의 연산, Time: 98ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240904/최대 2번의 연산/up-to-2-calculations.py
@@ -1,5 +1,5 @@
 a = int(input())
 if a%2 == 0:
-    print(a/2)
+    print(int(a/2))
 else:
-    print((a+1)/2)
+    print(int((a+1)/2))
